Move checks that use version-specific API
